requirements.txt
llama-index>=0.10.0
streamlit>=1.31.0
pinecone-client>=3.0.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
llama-index-llms-google-genai
llama-index-llms-groq
llama-index-embeddings-google-genai
llama-index-vector-stores-pinecone
PyMuPDF>=1.23.0